+++ po/POTFILES.in
@@ -143,6 +143,7 @@ imap/utf7.c
 imap/util.c
 index/config.c
 index/dlg_index.c
+index/functions.c
 index/ibar.c
 index/index.c
 index/ipanel.c
